+++ config/routes.rb
@@ -1,0 +1,3 @@
+Metricky::Engine.routes.draw do
+  resources :metrics, only: [:show], param: :name
+end